+++ app/model/request_cut_circuits.py
@@ -54,43 +54,3 @@ class CutCircuitsRequestSchema(ma.Schema):
     max_num_subcircuits = ma.fields.Int(required=False)
     subcircuit_vertices = ma.fields.List(ma.fields.List(ma.fields.Int), required=False)
     circuit_format = ma.fields.String(required=False)
-
-
-class CombineResultsRequest:
-    def __init__(
-        self,
-        circuit,
-        subcircuit_results,
-        cuts,
-        circuit_format="openqasm2",
-        unnormalized_results=False,
-        shot_scaling_factor=None,
-    ):
-        self.circuit = circuit
-        self.subcircuit_results = subcircuit_results
-        self.cuts = cuts
-        self.circuit_format = circuit_format
-        self.unnormalized_results = unnormalized_results
-        self.shot_scaling_factor = shot_scaling_factor
-
-
-class CombineResultsRequestSchema(ma.Schema):
-    circuit = ma.fields.Str(required=True)
-    subcircuit_results = ma.fields.List(
-        argschema.fields.NumpyArray(dtype=np.float), required=True
-    )
-    cuts = ma.fields.Dict(required=True)
-    circuit_format = ma.fields.String(required=False)
-    unnormalized_results = ma.fields.Boolean(required=False)
-    shot_scaling_factor = ma.fields.Int(required=False)
-
-
-class CombineResultsRequestQuokkaSchema(ma.Schema):
-    circuit = ma.fields.Str(required=True)
-    subcircuit_results = ma.fields.List(
-        ma.fields.Dict(keys=ma.fields.Str(), values=ma.fields.Float()), required=True
-    )
-    cuts = ma.fields.Dict(required=True)
-    circuit_format = ma.fields.String(required=False)
-    unnormalized_results = ma.fields.Boolean(required=False)
-    shot_scaling_factor = ma.fields.Int(required=False)
